+++ src/DIRAC/WorkloadManagementSystem/DB/PilotAgentsDB.sql
@@ -29,7 +29,6 @@ CREATE TABLE `PilotAgents` (
   `PilotID` INT(11) UNSIGNED NOT NULL AUTO_INCREMENT,
   `InitialJobID` INT(11) UNSIGNED NOT NULL DEFAULT 0,
   `CurrentJobID` INT(11) UNSIGNED NOT NULL DEFAULT 0,
-  `TaskQueueID` INT(11) UNSIGNED NOT NULL DEFAULT 0,
   `PilotJobReference` VARCHAR(255) NOT NULL DEFAULT 'Unknown',
   `PilotStamp` VARCHAR(32) NOT NULL DEFAULT '',
   `DestinationSite` VARCHAR(128) NOT NULL DEFAULT 'NotAssigned',
