+++ src/DIRAC/WorkloadManagementSystem/Utilities/QueueUtilities.py
@@ -11,7 +11,7 @@ from DIRAC.Core.Security.ProxyInfo import getProxyInfo
 from DIRAC.Resources.Computing.ComputingElementFactory import ComputingElementFactory
 
 
-def getQueuesResolved(siteDict, vo, queueCECache, checkPlatform=False, instantiateCEs=False):
+def getQueuesResolved(siteDict, queueCECache, checkPlatform=False, instantiateCEs=False):
     """Get the list of relevant CEs (what is in siteDict) and their descriptions.
     The main goal of this method is to return a dictionary of queues
     """
